app/backend/requirements.txt
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
django-prometheus==2.3.1